Project refactored

+++ Expense_Management_System.WebApi/Controllers/ExpenseDocumentController.cs
@@ -6,7 +6,6 @@ using Expense_Management_System.Application.Services;
 using Expense_Management_System.Domain.Entities;
 using Expense_Management_System.WebApi.ApiResponses;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Expense_Management_System.WebApi.Controllers;
@@ -18,7 +17,11 @@ public class ExpenseDocumentController : BaseController
 {
     private readonly IExpenseDocumentService _expenseDocumentService;
     private readonly IMapper _mapper;
-
+    public ExpenseDocumentController(IExpenseDocumentService expenseDocumentService, IMapper mapper)
+    {
+        _expenseDocumentService = expenseDocumentService;
+        _mapper = mapper;
+    }
 
     [HttpGet]
     public async Task<ApiResponse<IEnumerable<ExpenseDocumentResponse>>> GetAll()
@@ -54,24 +57,34 @@ public class ExpenseDocumentController : BaseController
     }
 
     [HttpPost]
-    public async Task<ApiResponse<ExpenseDocumentResponse>> Create([FromBody] ExpenseDocumentRequest documentRequest)
+    [Authorize(Roles = "Personnel,Admin")]
+    [Consumes("multipart/form-data")]
+    public async Task<ApiResponse<ExpenseDocumentResponse>> Create([FromForm] ExpenseDocumentRequest documentRequest)
     {
         if (documentRequest is null)
             return Fail<ExpenseDocumentResponse>("Invalid data", 400);
 
         var expenseDocument = _mapper.Map<ExpenseDocument>(documentRequest);
+        expenseDocument.Expenses ??= new Expense(); 
         expenseDocument.Expenses.UserId = CurrentUserId;
+        expenseDocument.InsertedUser = "system";
+        expenseDocument.InsertedDate = DateTime.UtcNow;
+        expenseDocument.IsActive = true;
+        expenseDocument.UploadDate = DateTime.Now;
 
         var createdDocument = await _expenseDocumentService.AddAsync(expenseDocument);
         var mappedEntity = _mapper.Map<ExpenseDocumentResponse>(createdDocument);
         return Success(mappedEntity, "Expense document successfully created.");
     }
 
+
     [HttpPut("{id}")]
-    public async Task<ApiResponse<ExpenseDocumentResponse>> Update(Guid id, [FromBody] ExpenseDocumentRequest documentRequest)
+    [Consumes("multipart/form-data")]
+    [Authorize(Roles = "Personnel,Admin")]
+    public async Task<ApiResponse<ExpenseDocumentResponse>> Update(Guid id, [FromForm] ExpenseDocumentRequest documentRequest)
     {
-        if (documentRequest is null)
-            return Fail<ExpenseDocumentResponse>("Invalid data", 400);
+        //if (documentRequest is null)
+        //    return Fail<ExpenseDocumentResponse>("Invalid data", 400);
 
         var existingDocument = await _expenseDocumentService.GetByIdAsync(id);
         if (existingDocument is null)
@@ -85,11 +98,16 @@ public class ExpenseDocumentController : BaseController
 
 
     [HttpDelete("{id}")]
+    [Authorize(Roles = "Admin")]
     public async Task<ApiResponse> Delete(Guid id)
     {
         var existingDocument = await _expenseDocumentService.GetByIdAsync(id);
         if (existingDocument is null)
             return Fail("Expense document not found", 404);
+
+        var role = CurrentUserRole;
+        if (role != "Admin")
+            return Fail("Access denied", 403);
 
         await _expenseDocumentService.DeleteAsync(existingDocument.Id);
         return Success("Expense document successfully deleted.");
